10/20

+++ Research/aging.h
@@ -5,7 +5,7 @@
 //DCC本身的delay
 //50%的DCC(就只是個buffer)
 
-double AgingRate(AGINGTYPE status, int year);
+double AgingRate(AGINGTYPE status, double year);
 void ReadAgingData();
 
 #endif
